+++ src/widgets/SpecialTag/SpecialTag.tsx
@@ -4,11 +4,28 @@ import { SpecialType } from "../../data/special";
 import "./SpecialTag.css";
 
 export interface SpecialTag {
+    
+    /**
+     *  The special trait to show.
+     */
     special: SpecialType;
+
+    /**
+     *  Should the special tag be redered in an expanded way? Meaning
+     *  that the description of the special trait would be always visible.
+     */
+    expanded?: boolean;
+
+    /**
+     *  A callback called when user click the tag.
+     */
     onClick?: (special: SpecialType) => void;
 };
 
-export function SpecialTag({ special, onClick } : SpecialTag) {
+/**
+ *  This is a component that shows information about a special trait.
+ */
+export function SpecialTag({ special, expanded, onClick } : SpecialTag) {
 
     const { t } = useTranslation();
 
@@ -19,9 +36,14 @@ export function SpecialTag({ special, onClick } : SpecialTag) {
     const css = [ "specialtag" ];
     if (onClick) css.push("spacialtag-clickable");
 
+    // when the expanded is not provided then we want to still give the hint
+    // about what is the spcial trait in a tooltip.
+    const title = !expanded ? String(t(`special.description.${special}`)) : "";
+
     return (
-        <span className={css.join(" ")} title={String(t(`special.description.${special}`))} onClick={handleClick}>
+        <span className={css.join(" ")} title={title} onClick={handleClick}>
             {t(`special.label.${special}`)}
+            {expanded && (<div className="specialtag-description"> - {String(t(`special.description.${special}`))}</div>)}
         </span>
     );
 };